feat: add user endpoint authorization checks
Implemented user endpoint authorization:
- Added current user dependency to user endpoints
- Enforced access control for get, update, and delete operations
- Restricted users to only access their own information
- Introduced access denied exception for unauthorized requests

+++ splice/interface/routers/user_router.py
@@ -10,29 +10,41 @@ from splice.core.models.user import (
 )
 from splice.infra.database import get_pg_session
 from splice.infra.repositories.user_repository import UserRepository
+from splice.interface.service.auth_service import get_current_user
 from splice.interface.service.user_service import UserService
 
 router = APIRouter(prefix='/users')
 
+acesso_negado = HTTPException(status_code=403, detail='Acesso negado')
+
 
 @router.get('', response_model=UserResponse)
-async def get_user(
+async def get(
     user_id: str = None,
     username: str = None,
     email: str = None,
     phone: str = None,
     db_session: Session = Depends(get_pg_session),
+    current_user: User = Depends(get_current_user),
 ):
     repo = UserRepository(db_session)
     service = UserService(repo)
 
     if user_id:
+        if user_id != current_user.id:
+            raise acesso_negado
         usuario = await service.get_user_by_id(user_id)
     elif username:
+        if username != current_user.username:
+            raise acesso_negado
         usuario = await service.get_user_by_username(username)
     elif email:
+        if email != current_user.email:
+            raise acesso_negado
         usuario = await service.get_user_by_email(email)
     elif phone:
+        if phone != current_user.phone:
+            raise acesso_negado
         usuario = await service.get_user_by_phone(phone)
     else:
         raise HTTPException(
@@ -41,12 +53,11 @@ async def get_user(
 
     if not usuario:
         raise HTTPException(status_code=404, detail='Usuário não encontrado')
-    print(usuario.establishment)
     return usuario
 
 
 @router.post('', response_model=User)
-async def create_user(
+async def create(
     data: UserCreateSchema = Body(),  # type: ignore
     db_session: Session = Depends(get_pg_session),
 ):
@@ -57,9 +68,10 @@ async def create_user(
 
 
 @router.put('', response_model=User)
-async def update_user(
+async def update(
     data: UserUpdateSchema = Body(),  # type: ignore
     db_session: Session = Depends(get_pg_session),
+    current_user: User = Depends(get_current_user),
 ):
     repo = UserRepository(db_session)
     service = UserService(repo)
@@ -67,12 +79,20 @@ async def update_user(
     # Converte o body em dicionário, removendo campos nulos
     update_data = data.model_dump(exclude_unset=True)
 
+    if data.id != current_user.id:
+        raise acesso_negado
     # Passa os dados descompactados para a função de atualização
     return await service.update_user(user_id=data.id, **update_data)
 
 
 @router.delete('')
-async def delete_user(user_id: str, db_session=Depends(get_pg_session)):
+async def delete(
+    user_id: str,
+    db_session=Depends(get_pg_session),
+    current_user: User = Depends(get_current_user),
+):
     repo = UserRepository(db_session)
     service = UserService(repo)
+    if user_id != current_user.id:
+        raise acesso_negado
     return await service.delete_user(user_id)
